+++ packages/health/android/src/main/kotlin/cachet/plugins/health/HealthPlugin.kt
@@ -3002,7 +3002,7 @@ class HealthPlugin(private var channel: MethodChannel? = null) :
                                         }
                         is BodyTemperatureRecord ->
                                         return listOf(
-                                                        mapOf<String, Any>(
+                                                        mapOf<String, Any>(w
                                                                         "value" to
                                                                                         record.temperature
                                                                                                         .inCelsius,